Second unit test created, with CSRF protection disabled during tests
