+++ app/config.py
@@ -13,3 +13,4 @@ class DeploymentConfig(Config):
 class TestConfig(Config):
     SQLALCHEMY_DATABASE_URI = "sqlite:///:memory"
     TESTING = True
+    WTF_CSRF_ENABLED = False                  # Disable CSRF protection during tests
